# rigweda/backend/src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rigweda API
  version: 0.8.0
  description: Independently implemented HR and payroll platform API.
servers:
  - url: /api/v1
paths:
...
  /employees:
    get:
      tags: [Employees]
      summary: Search and filter the organization employee directory
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [preboarding, probation, active, on_leave, notice_period, suspended, terminated, all] } }
        - { name: employmentType, in: query, schema: { type: string, enum: [full_time, part_time, contract, intern, temporary, apprentice, all] } }
        - { name: departmentId, in: query, schema: { type: string, format: uuid } }
        - { name: workLocationId, in: query, schema: { type: string, format: uuid } }
        - { name: managerEmployeeId, in: query, schema: { type: string, format: uuid } }
        - { name: sort, in: query, schema: { type: string, enum: [name, employeeNumber, joinDate, status] } }
        - { name: direction, in: query, schema: { type: string, enum: [asc, desc] } }
      responses:
        "200": { description: Paginated employee directory }
    post:
      tags: [Employees]
      summary: Create an employee profile and initial lifecycle history
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeInput" }
      responses:
        "201": { description: Employee created }
        "409": { description: Employee number, work email, or linked user conflict }
        "422": { description: Validation or organization-reference error }
  /employees/summary:
    get:
      tags: [Employees]
      summary: Retrieve employee lifecycle totals
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Employee summary counts }
  /employees/metadata:
    get:
      tags: [Employees]
      summary: Retrieve departments, job titles, locations, managers, and available login accounts
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Employee form metadata }
  /employees/export:
    get:
      tags: [Employees]
      summary: Export the authorized employee directory as CSV
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: UTF-8 CSV export
          content: { text/csv: { schema: { type: string } } }
        "403": { description: Employee-export permission required }
  /employees/next-number:
    get:
      tags: [Employees]
      summary: Suggest the next available employee number
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Suggested organization-scoped employee number }
  /employees/upcoming-events:
    get:
      tags: [Employees]
      summary: List upcoming birthdays, anniversaries, and probation endings
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: days, in: query, schema: { type: integer, minimum: 1, maximum: 365, default: 45 } }
      responses:
        "200": { description: Upcoming employee events }
  /employees/organization-tree:
    get:
      tags: [Employees]
      summary: Retrieve the manager and direct-report hierarchy
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Organization reporting tree }
  /employees/archived:
    get:
      tags: [Employees]
      summary: List reversibly archived employee records
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Paginated archived employees }
        "403": { description: Employee-lifecycle permission required }
  /employees/me:
    get:
      tags: [Employees]
      summary: Retrieve the employee profile linked to the current login
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Current employee profile }
        "404": { description: Login has no linked employee profile }
  /employees/me/profile:
    put:
      tags: [Employees]
      summary: Complete or update the current employee's self-service profile
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Self-service profile saved }
        "409": { description: Optimistic version conflict }
  /employees/bulk-update:
    put:
      tags: [Employees]
      summary: Assign department, job title, location, manager, mode, or type in bulk
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Number of employee records updated }
        "403": { description: Employee-bulk permission required }
  /employees/{employeeId}:
    parameters:
      - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Employees]
      summary: Retrieve a private employee profile
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Employee profile with addresses and emergency contacts }
        "403": { description: Sensitive-employee permission required }
        "404": { description: Employee not found }
    patch:
      tags: [Employees]
      summary: Update employee identity, employment assignment, or contact data
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Employee updated }
        "409": { description: Optimistic version conflict }
        "422": { description: Invalid reference, reporting cycle, or dates }
  /employees/{employeeId}/status:
    post:
      tags: [Employees]
      summary: Apply a validated employee lifecycle transition
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Employee status changed and recorded in history }
        "422": { description: Invalid lifecycle transition }
  /employees/{employeeId}/history:
    get:
      tags: [Employees]
      summary: Retrieve job, lifecycle, and change history
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Employee history }
        "404": { description: Employee not found }
  /employees/{employeeId}/archive:
    post:
      tags: [Employees]
      summary: Reversibly archive an employee and deactivate their membership
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Employee archived }
        "409": { description: Optimistic version conflict }
  /employees/{employeeId}/restore:
    post:
      tags: [Employees]
      summary: Restore an archived employee and reactivate their membership
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Employee restored }
        "409": { description: Optimistic version conflict }
  /employees/{employeeId}/reopen-profile:
    post:
      tags: [Employees]
      summary: Reopen a completed employee profile for correction
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Profile reopened }
  /employees/{employeeId}/sensitive-records:
    put:
      tags: [Employees]
      summary: Replace protected identifiers and document metadata
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Protected records replaced; identifier values remain masked in responses }
        "403": { description: Employee-documents permission required }
  /employees/{employeeId}/identifiers:
    post:
      tags: [Employees]
      summary: Encrypt and add a protected employee identifier
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "201": { description: Identifier stored with AES-256-GCM encryption }
  /employees/{employeeId}/identifiers/{recordId}:
    delete:
      tags: [Employees]
      summary: Remove a protected employee identifier
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: recordId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Identifier removed }
  /employees/{employeeId}/documents:
    post:
      tags: [Employees]
      summary: Add protected employee document metadata and storage URL
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "201": { description: Document metadata added }
  /employees/{employeeId}/documents/{recordId}:
    delete:
      tags: [Employees]
      summary: Remove protected employee document metadata
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: employeeId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: recordId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Document metadata removed }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EmployeeInput:
      type: object
      required: [employeeNumber, firstName, lastName, employmentType, joinDate]
      properties:
        employeeNumber: { type: string, pattern: "^[A-Z0-9][A-Z0-9_/-]{1,39}$" }
        firstName: { type: string, minLength: 1, maxLength: 100 }
        middleName: { type: [string, "null"], maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        preferredName: { type: [string, "null"], maxLength: 100 }
        userId: { type: [string, "null"], format: uuid }
        workEmail: { type: [string, "null"], format: email }
        personalEmail: { type: [string, "null"], format: email }
        workPhone: { type: [string, "null"] }
        personalPhone: { type: [string, "null"] }
        dateOfBirth: { type: [string, "null"], format: date }
        employmentStatus: { type: string, enum: [preboarding, probation, active] }
        employmentType: { type: string, enum: [full_time, part_time, contract, intern, temporary, apprentice] }
        workMode: { type: string, enum: [onsite, hybrid, remote] }
        joinDate: { type: string, format: date }
        departmentId: { type: [string, "null"], format: uuid }
        jobTitleId: { type: [string, "null"], format: uuid }
        workLocationId: { type: [string, "null"], format: uuid }
        managerEmployeeId: { type: [string, "null"], format: uuid }
        addresses: { type: array, maxItems: 3, items: { type: object } }
        emergencyContacts: { type: array, maxItems: 10, items: { type: object } }
...